+++ include/vsg/vk/BufferData.h
@@ -12,7 +12,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 </editor-fold> */
 
-#include <vsg/vk/Buffer.h>
+#include <vsg/state/Buffer.h>
 
 #include <cstring>
