+++ src/vsg/state/Buffer.cpp
@@ -12,7 +12,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 #include <vsg/core/Exception.h>
 #include <vsg/io/Options.h>
-#include <vsg/vk/Buffer.h>
+#include <vsg/state/Buffer.h>
 #include <vsg/vk/Context.h>
 
 #include <iostream>
