+++ src/vsg/vk/DeviceMemory.cpp
@@ -12,7 +12,6 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 #include <vsg/core/Exception.h>
 #include <vsg/io/Options.h>
-#include <vsg/vk/Buffer.h>
 #include <vsg/vk/DeviceMemory.h>
 
 #include <atomic>
